Added CalibratedClassifierCV(SVC)
SVC(kernel='linear') has a predict_proba method but its ouput is sometimes inconsistent with predict, which prevents AutoClassifier to pass check_estimator. This issue is already reported here: https://github.com/scikit-learn/scikit-learn/issues/13211 and it seems that is will be solved soon. For now I added CalibratedClassifierCV which is what they recommend in the previous link.